单例方式初始化支付宝小程序

+++ handler/src/main/java/com/zz/messagepush/handler/domain/wechat/WeChatOfficialParam.java
@@ -39,5 +39,5 @@ public class WeChatOfficialParam {
     /**
      * 发送账户
      */
-    private Integer sendAccount;
+    private Long sendAccount;
 }
